Get session_id from post request response
