+++ app/src/main/java/fr/isen/racketselectorapp/model/process/StartFragment.kt
@@ -1,4 +1,4 @@
-package fr.isen.racketselectorapp.process
+package fr.isen.racketselectorapp.model.process
 
 import android.content.Context
 import android.os.Bundle
@@ -6,11 +6,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import fr.isen.racketselectorapp.ProcessActivityInteraction
-import fr.isen.racketselectorapp.databinding.FragmentForehandBinding
+import fr.isen.racketselectorapp.databinding.FragmentStartBinding
 
-class ForehandFragment : Fragment() {
-    private lateinit var binding: FragmentForehandBinding
+class StartFragment : Fragment() {
+    private lateinit var binding: FragmentStartBinding
     private lateinit var interactor: ProcessActivityInteraction
 
     override fun onAttach(context: Context) {
@@ -22,15 +21,15 @@ class ForehandFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentForehandBinding.inflate(inflater, container, false)
+        binding = FragmentStartBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.nextForehandButton.setOnClickListener {
-            interactor.showNextStep(SmashFragment())
+        binding.beginStartButton.setOnClickListener {
+            interactor.showNextStep(ServeFragment())
         }
     }
 }
