+++ app/src/main/res/layout/activity_form.xml
@@ -32,6 +32,7 @@
         app:layout_constraintVertical_bias="1.0" />
 
     <ScrollView
+        android:id="@+id/scrollView2"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginTop="32dp"
@@ -211,9 +212,10 @@
                     android:backgroundTint="@color/transparent"
                     android:hint="@string/weight"
                     android:inputType="number"
-                    tools:ignore="TextContrastCheck"/>
+                    tools:ignore="TextContrastCheck" />
             </com.google.android.material.textfield.TextInputLayout>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
     </ScrollView>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
