+++ app/src/main/res/layout/activity_recap.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecapActivity">
+    tools:context=".model.RecapActivity">
 
     <TextView
         style="@style/title"
