+++ app/src/main/res/layout/fragment_backhand.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".process.BackhandFragment">
+    tools:context=".model.process.BackhandFragment">
 
     <Button
         android:id="@+id/finishBackhandButton"
