+++ app/src/main/res/layout/fragment_forehand.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".process.ForehandFragment">
+    tools:context=".model.process.ForehandFragment">
 
     <TextView
         android:id="@+id/forehandText"
