+++ app/src/main/res/layout/fragment_serve.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".process.ServeFragment">
+    tools:context=".model.process.ServeFragment">
 
     <TextView
         android:id="@+id/serveText"
