+++ app/src/main/res/layout/fragment_smash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".process.SmashFragment">
+    tools:context=".model.process.SmashFragment">
 
     <Button
         android:id="@+id/nextSmashButton"
