+++ app/src/main/res/layout/fragment_start.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".process.StartFragment">
+    tools:context=".model.process.StartFragment">
 
     <TextView
         android:id="@+id/startText"
