+++ app/src/main/res/layout/splash_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashScreen">
+    tools:context=".splashscreen.SplashScreen">
 
     <ImageView
         android:id="@+id/humanFabLogo"
